Fixed tests for unified code assistant
