+++ tests/fixtures/core_fixtures.py
@@ -8,6 +8,7 @@ import pytest
 import numpy as np
 from pathlib import Path
 from typing import Any
+from unittest.mock import MagicMock
 
 
 # ===========================
@@ -99,3 +100,4 @@ def stub_indexers(
         "scripts.indexers.summary_indexer.SummaryIndexer.__init__", mock_init_summary
     )
     monkeypatch.setattr("scripts.indexers.raw_log_indexer.RawLogIndexer.__init__", mock_init_raw)
+
