+++ tests/unit/unified_code_assistant/test_assistant_utils.py
