+++ tests/unit/unified_code_assistant/test_module_summarizer.py
@@ -3,7 +3,24 @@ import json
 import tempfile
 import os
 from scripts.unified_code_assistant.assistant_utils import load_report, extract_code_snippets, get_issue_locations
-from scripts.unified_code_assistant import build_contextual_prompt, build_enhanced_contextual_prompt
+from scripts.unified_code_assistant.prompt_builder import build_contextual_prompt, build_enhanced_contextual_prompt
+from scripts.unified_code_assistant.module_summarizer import summarize_modules
+from scripts.config.config_manager import ConfigManager
+import pytest
+from types import SimpleNamespace
+
+@pytest.fixture(autouse=True)
+def patch_mock_config(monkeypatch):
+    mock_config = SimpleNamespace(
+        persona="test",
+        prompts_by_subcategory={"_default": "Summarize the function briefly."}
+    )
+    monkeypatch.setattr("scripts.config.config_manager.ConfigManager.load_config", lambda: mock_config)
+
+
+class MockSummarizer:
+    def summarize_entry(self, text, subcategory=None):
+        return "Summary for: " + text[:30]
 
 def test_load_report():
     mock_data = {"some": "data"}
@@ -98,3 +115,27 @@ def test_build_enhanced_contextual_prompt():
     assert "module1.py" in prompt
     assert "Refactor large function." in prompt
     assert "What should we fix first?" in prompt
+
+def test_summarize_modules():
+    report_data = {
+        "src/module1.py": {
+            "docstrings": {
+                "functions": [{"name": "foo", "docstring": "Does something"}]
+            }
+        },
+        "src/module2.py": {
+            "docstrings": {
+                "functions": []
+            }
+        },
+        "src/ignore.py": {
+            "docstrings": {}
+        }
+    }
+    config = ConfigManager.load_config()
+    summaries = summarize_modules(report_data, MockSummarizer(), config=config)
+    assert "src/module1.py" in summaries
+    assert "Summary for: " in summaries["src/module1.py"]
+    assert summaries["src/module2.py"] == "Docstrings exist but no functions parsed."
+    assert "src/ignore.py" not in summaries
+
